// src/enums/user-role.enum.ts
export enum UserRole {
  CEO = "ceo",
  GENERAL_MANAGER = "general_manager",
  MARKETING_LEAD = "marketing_lead",
  FINANCE_OFFICER = "finance_officer",
  DATA_COLLECTOR = "data_collector",
  QUANTITY_SURVEYOR = "quantity_surveyor",
  DESIGNER = "designer",
  SITE_ENGINEER = "site_engineer",
  SYSTEM_ADMINISTRATOR = "system_administrator",
}